backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
elasticsearch==8.9.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.25.2
openai==1.3.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
redis==5.0.1
websockets==12.0
aiokafka==0.8.11
aiohttp==3.9.0